Finally got the error system working.
Took way longer than I expected! As always... Such is the life of a programmer :/ Cracked open a peer to celebrate :D

Basically, when netvend detects an error that was the fault of the agent while processing an agent command (for example, the agent specifies an invalid target or tries to send money from a pocket he does not own), it creats a child of commands::errors::Error. Then the server throws a NetvendCommandException, which *contains* a shared pointer to the commands::errors::Error. Higher up in the server (in processCommandBatchPacket), it catches this exception and extracts the commands::errors::Error. It adds this to the commands::results::Batch just like it would a regular result, then either terminates the batch or continues, depending on the Error's fatalToBatch bool variable.

You'll see that now, client.cpp's Agent::performSingleCommand method checks each Result from the results::Batch for its error member to see if it's an error instead of the commands::results::Result type it expected. If the error member is nonzero, it dynamic_casts it to a commands::errors::Error, and throws it.

Client.cpp is really a placeholder for now, which is going to be transformed into an api once the server is more complete. So the idea is that the api will throw a commands::errors::Error, and the developer can catch it and extract more information from it.

Niiiiiiiiiiice real nice