Removed parameter requestingDepsoitAddress from createPocket command.
Doesn't really make sense to have that parameter there. New agents would have to separately request a deposit address for their default pocket anyway, or ignore the default pocket and create a new one with this parameter set to true.

In the future a command in a commandBatch will be able to reference a result from an earlier command in the same batch, which will allow the agent to create a pocket and get a deposit address for it within the same batch.