Changed, again the name of what used to be called 'chunks' or 'pages', to simply 'files'.
This really makes the most sense. They're discrete pieces of binary data, meant for any arbitrary purpose. That's what a file is.

An agent could upload something.html meant for reading, or image files, or anything. The term files makes sense, and implies the full possibility of the read/write commands.